Change theme and adjust "is-graphical" logic
The correct theme will not be applied in emacsclient mode due to
the fact that the frame has not been set up when display-graphic-p
is called. If daemon mode is turned on, the setup-theme function
must be added as a hook to after-frame-functions, and the frame must
be selected before the display-graphic-p is called.